🚚 Rename test files
